Likes et dislike fonctionnels
probleme unlike image
probleme reactivité des pages
probleme unshift posts

+++ back/controllers/user.js
@@ -54,7 +54,7 @@ exports.all = async (req, res, next) => {
     }
 }
 
-exports.oneUser = async (req, res, next) => {
+exports.oneUserProfile = async (req, res, next) => {
     try{
         const id = req.user.payload.id
     const oneUser = await prisma.user.findUnique({
@@ -74,21 +74,45 @@ exports.oneUser = async (req, res, next) => {
         next(createHttpError(e.statusCode, e.message))
     }
 }
-
-exports.deleteUser = async (req, res, next) => {
-    try {
-        const id = req.user.payload.id
-        const user = await prisma.user.delete({
-            where: {
-                id: Number(id)
-            },
-        })
-        res.status(201).json({
+exports.oneUser = async (req, res, next) => {
+    try{
+        const id = req.params
+    const oneUser = await prisma.user.findUnique({
+        where: {
+            id: Number(id)
+        },
+        include:{
+            profile: true
+        }
+    })
+        res.status(200).json({
             status: true,
-            message: "Compte bien supprimé",
-            data: user
+            message: 'One user',
+            data: oneUser
         })
-    } catch (e) {
+    } catch(e){
         next(createHttpError(e.statusCode, e.message))
     }
+}
+
+exports.deleteUser = async (req, res, next) => {
+    const id = req.user.payload.id
+    if(id === req.user.payload.id || req.user.payload.isAdmin === 1){
+        try {
+            const id = req.user.payload.id
+            const user = await prisma.user.delete({
+                where: {
+                    id: Number(id)
+                },
+            })
+            res.status(201).json({
+                status: true,
+                message: "Compte bien supprimé",
+                data: user
+            })
+        } catch (e) {
+            next(createHttpError(e.statusCode, e.message))
+        }
+    }
+    
 }
